feat: add config for auto-pairs
Using GOAT auto-pairs from jiangmiao written in vimscript instead of lua
copies of it. This is lightweight and no-nonsense, in my opinion. I
already use 2 of keymappings this plugin sets up (<ALT-n> and <ALT-p>)
for cycle_themes.lua for switching between colorschemes back and forth.
Therefore, these 2 keymappings are disabled for auto-pairs.